BlogGenerator/requirements.txt
Flask==3.0.0
openai>=2.0.0
python-dotenv==1.0.0
markdown2==2.4.10
Jinja2==3.1.2

